유저랭크 정보 렌더링하는 부분 컴포넌트로 추출 및 테스트 코드 작성
- 유저랭크 정보 렌더링하는 부분 컴포넌트로 추출
- 랭크가 없는 경우도 추가
- player 테스트 커버리지 100% 달성
